+++ src/builds/web/app.vue
@@ -1,5 +1,6 @@
 <template>
   <div id="app">
+    <logout-warning-modal ref="logoutWarningModal" />
     <header-container v-show="$route.fullPath !== '/getting-started'" />
     <welcome-modal ref="welcome" />
     <router-view />
@@ -16,6 +17,7 @@ import WelcomeModal from '@/components/WelcomeModal';
 import store from 'store';
 import { mapGetters } from 'vuex';
 import { Toast } from '@/helpers';
+import LogoutWarningModal from '@/components/LogoutWarningModal';
 
 export default {
   name: 'App',
@@ -23,12 +25,24 @@ export default {
     'header-container': HeaderContainer,
     'footer-container': FooterContainer,
     'confirmation-container': ConfirmationContainer,
+    'logout-warning-modal': LogoutWarningModal,
     'welcome-modal': WelcomeModal
   },
   computed: {
     ...mapGetters({
       wallet: 'wallet'
     })
+  },
+  watch: {
+    $route(to, from) {
+      if (
+        from.matched[0].path === '/interface' &&
+        to.matched[0].path !== '/interface'
+      ) {
+        // Show logout warning modal
+        this.$refs.logoutWarningModal.$refs.logoutWarningModal.show();
+      }
+    }
   },
   created() {
     const succMsg =